Limit calls to Dir.chdir in Rake::Application
Ruby [Bug #15661] changed the blockless form of `Dir.chdir` to raise an
error if it is called during the block of another `Dir.chdir` call.
This commit refactors code in `Rake::Application` to limit calls to
`Dir.chdir`, especially blockless calls.  The eventual goal is to make
it safe to call `Rake::Application#run` from within a `Dir.chdir` block.

[Bug #15661]: https://bugs.ruby-lang.org/issues/15661